refactor: split `compile.ts` file to several files in `compile` dir
