+++ src/test/cpp/generate/index.test.ts
@@ -1,7 +1,7 @@
 import { createTempDirectory, ITempDirectory } from "create-temp-directory";
 import fs from "node:fs/promises";
 import path from "node:path";
-import { compileCppSource } from "../../../compile.js";
+import { compileCppSource } from "../../../compile/cpp.js";
 import { runExecutable } from "../../../run.js";
 import { Schema } from "../../schema.js";
 import { generateCppTest } from "./index.js";
